Only save in-game configs to file when necessary
Configs will now only be saved to file when in-game changes were made using the /sc config command.

This aims to make it far less likely for someone to modify the config file, forget to reload the config ingame and get confused as to why the settings went back to default after a restart.

Also improved error handling a bit.