Replace mobcap setting `mobcap-modifier` with `additional-capacity`
Instead of modifying the mobcap by a modifier additional-capacity will simply add a constant value to it.

Reason for this being that the modifier becomes very inconsistent to use when the mobcap percentage gets modified to anything not 100%.

If the mobcap would be 10% this would basically mean that at a modifier of 1.5x spawners and such would almost never get a chance to spawn as it only gets 3 mobs headroom.